+++ src/main/java/simon/SimonException.java
@@ -1,3 +1,5 @@
+package simon;
+
 class SimonException extends Exception {
     public SimonException(String message) {
         super(message);
