+++ src/main/java/simon/Storage.java
@@ -1,3 +1,5 @@
+package simon;
+
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
