+++ src/main/java/simon/TaskList.java
@@ -1,3 +1,5 @@
+package simon;
+
 import java.util.ArrayList;
 
 public class TaskList {
