+++ src/main/java/simon/Ui.java
@@ -1,3 +1,5 @@
+package simon;
+
 import java.util.Scanner;
 
 public class Ui {
